// src/app/utilities/app-enum.ts
export enum UserTaskStatus {
    OPENING = 0,// Đang mở
    DOING = 1,// Đang tiến hành
    PAUSE = 2,// Tạm hoãn
    COMPLETE = 3,// Hoàn thành
}

export enum UserTaskRole {
    RESPONSIBLE = 1,// Người chịu trách nhiệm
    JOINED = 2,// Người tham gia
    VIEWER = 3// Người quan sát
}

export enum LanguageType {
    ENGLISH = 1,
    VIETNAM = 2,
    KOREA = 3
}